Safari-609.1.11.1/LayoutTests/http/tests/security/window-name-after-same-origin-sub-frame-navigation-expected.txt

CONSOLE MESSAGE: line 9: window.name should equal "test" after a same-origin sub frame navigation.
CONSOLE MESSAGE: line 4: window.name equals "test".
